Put integration-counting code in the right place
